+++ python/sglang/kernels/ops/attention/linear/__init__.py
@@ -1,0 +1,1 @@
+"""Linear-attention kernels (RFC #29630, Phase 2.5)."""
