+++ python/sglang/kernels/ops/attention/minimax_sparse/__init__.py
@@ -1,0 +1,1 @@
+"""MiniMax sparse-attention kernels (RFC #29630, Phase 2.5)."""
