+++ python/sglang/kernels/ops/diffusion/render/__init__.py
