+++ python/sglang/kernels/ops/diffusion/render/hunyuan3d_rasterizer/rasterizer.h
