+++ python/sglang/kernels/ops/diffusion/render/mesh_processor/__init__.py
@@ -12,7 +12,8 @@ import os
 from typing import Tuple
 
 import numpy as np
-from sglang.multimodal_gen.csrc.render import load_extension_with_recovery
+
+from sglang.kernels.ops.diffusion.render import load_extension_with_recovery
 
 _abs_path = os.path.dirname(os.path.abspath(__file__))
 _mesh_processor_kernel = None
@@ -55,7 +56,9 @@ def meshVerticeInpaint(
     pos_idx = np.ascontiguousarray(pos_idx, dtype=np.int32)
     uv_idx = np.ascontiguousarray(uv_idx, dtype=np.int32)
 
-    return kernel.meshVerticeInpaint(texture, mask, vtx_pos, vtx_uv, pos_idx, uv_idx, method)
+    return kernel.meshVerticeInpaint(
+        texture, mask, vtx_pos, vtx_uv, pos_idx, uv_idx, method
+    )
 
 
 __all__ = ["meshVerticeInpaint"]
