+++ python/sglang/kernels/ops/diffusion/render/mesh_processor/mesh_processor.cpp
