+++ python/sglang/srt/layers/attention/minimax_sparse_ops/tests/test_flash_with_topk_idx.py
@@ -3,10 +3,10 @@ import sys
 import pytest
 import torch
 
-from sglang.srt.environ import envs
-from sglang.srt.layers.attention.minimax_sparse_ops.decode.flash_with_topk_idx import (
+from sglang.kernels.ops.attention.minimax_sparse.decode.flash_with_topk_idx import (
     flash_decode_with_topk_idx,
 )
+from sglang.srt.environ import envs
 
 DEVICE = "cuda"
 RTOL_VS_REF = 5e-3
