+++ test/registered/attention/test_kda_prefill_cutedsl.py
@@ -25,15 +25,15 @@ if not (torch.cuda.is_available() and torch.cuda.get_device_capability()[0] >= 1
         allow_module_level=True,
     )
 
+from sglang.kernels.ops.attention.linear.kda_blackwell import (  # noqa: E402
+    chunk_kda_cutedsl,
+    prepare_metadata,
+)
 from sglang.srt.layers.attention.fla.index import (  # noqa: E402
     prepare_chunk_indices,
     prepare_chunk_offsets,
 )
 from sglang.srt.layers.attention.fla.kda import fused_recurrent_kda  # noqa: E402
-from sglang.srt.layers.attention.linear.kernels.kda_blackwell import (  # noqa: E402
-    chunk_kda_cutedsl,
-    prepare_metadata,
-)
 
 
 def _l2norm(x: torch.Tensor) -> torch.Tensor:
