+++ test/registered/jit/benchmark/minimax/bench_minimax_decode_topk.py
@@ -11,7 +11,7 @@ import triton
 
 from sglang.jit_kernel.benchmark import marker
 from sglang.jit_kernel.minimax_decode_topk import minimax_decode_topk
-from sglang.srt.layers.attention.minimax_sparse_ops.decode.flash_with_topk_idx import (
+from sglang.kernels.ops.attention.minimax_sparse.decode.flash_with_topk_idx import (
     _topk_index_merge_kernel,
     _topk_index_partial_kernel,
 )
